+++ packages/solid-crs-components/lib/index.ts
@@ -23,3 +23,4 @@ export * from './state/event';
 export * from './state/schema';
 export * from './state/state';
 export * from './header/content-header.component';
+export * from './loading/progress-bar-component';