+++ packages/solid-crs-manage/lib/app-root.component.ts
@@ -7,7 +7,7 @@ import { Alert, FormActors, FormEvent } from '@netwerk-digitaal-erfgoed/solid-cr
 import { RxLitElement } from 'rx-lit';
 import { Theme, Logout, Logo, Plus, Cross, Search } from '@netwerk-digitaal-erfgoed/solid-crs-theme';
 import { unsafeSVG } from 'lit-html/directives/unsafe-svg';
-import { AppActors, AppAuthenticateStates, AppContext, AppFeatureStates, appMachine, AppRootStates } from './app.machine';
+import { AppActors, AppAuthenticateStates, AppContext, AppDataStates, AppFeatureStates, appMachine, AppRootStates } from './app.machine';
 import nlNL from './i8n/nl-NL.json';
 import { AppEvents } from './app.events';
 import { SolidSDKService } from './common/solid/solid-sdk.service';
@@ -72,6 +72,10 @@ export class AppRootComponent extends RxLitElement {
         identifier: this.translator.translate('nde.features.object.new-object-name').toLowerCase().replace(' ', '-'),
         image: 'https://images.unsplash.com/photo-1615390164801-cf2e70f32b53?ixid=MnwxMjA3fDB8MHxwcm9maWxlLXBhZ2V8M3x8fGVufDB8fHx8&ixlib=rb-1.2.1&w=1000&q=80',
         license: 'https://creativecommons.org/publicdomain/zero/1.0/deed.nl',
+        height: 0,
+        width: 0,
+        depth: 0,
+        weight: 0,
       }
     )).withContext({
       alerts: [],
@@ -227,8 +231,16 @@ export class AppRootComponent extends RxLitElement {
    */
   render(): TemplateResult {
 
+    const showLoading = this.state?.matches({ [AppRootStates.DATA]: AppDataStates.CREATING })
+      || this.state?.matches({ [AppRootStates.DATA]: AppDataStates.REFRESHING });
+
     return html`
+
+
     ${ this.state?.matches({ [AppRootStates.AUTHENTICATE]: AppAuthenticateStates.AUTHENTICATED }) ? html`
+
+    ${ showLoading ? html`<nde-progress-bar></nde-progress-bar>` : html``}
+
     <nde-sidebar inverse>
       <nde-content-header>
         <div slot="icon">${ unsafeSVG(Logo) }</div>
@@ -293,7 +305,12 @@ export class AppRootComponent extends RxLitElement {
         :host > *:not(nde-sidebar) {
           flex: 1 1;
         }
-
+        nde-progress-bar {
+          position: absolute;
+          width: 100%;
+          top: 0;
+          left: 0;
+        }
         nde-sidebar {
           width: var(--size-sidebar);
         }