+++ packages/solid-crs-manage/lib/features/authenticate/authenticate-root.component.spec.ts
@@ -1,4 +1,4 @@
-import { Alert, FormEvents } from '@netwerk-digitaal-erfgoed/solid-crs-components';
+import { Alert } from '@netwerk-digitaal-erfgoed/solid-crs-components';
 import { ArgumentError, Collection, CollectionMemoryStore, CollectionObject, CollectionObjectMemoryStore, ConsoleLogger, LoggerLevel } from '@netwerk-digitaal-erfgoed/solid-crs-core';
 import { interpret, Interpreter } from 'xstate';
 import { AppEvents, DismissAlertEvent, LoggedInEvent } from '../../app.events';